Create the base of the new footer
The footer is not ready yet. I would like to design better the line
above the footer and the one at the bottom of the header to have the
same width that the body. Also, I'll try to recreate the spinner at
the footer from the master branch. As an idea, I would like to play
with an sticky logo like the one from [1]

[1]: https://briangardner.com/

Signed-off-by: Antonio Ossa <aaossa@uc.cl>